fix error.
converted back to double and added the ExampleInstrumentedTest.java back

+++ app/src/androidTest/java/minicap/concordia/campusnav/MainActivityEndToEndTest.java
@@ -48,8 +48,8 @@ public class MainActivityEndToEndTest {
         assertEquals("minicap.concordia.campusnav.screens.MapsActivity", actualIntent.getComponent().getClassName());
 
         // Check that the extras contain the correct latitude and longitude using TOLERANCE AS PRECISION HERE
-        float startingLat = actualIntent.getFloatExtra(MapsActivity.KEY_STARTING_LAT, 0.0f);
-        float startingLng = actualIntent.getFloatExtra(MapsActivity.KEY_STARTING_LNG, 0.0f);
+        double startingLat = actualIntent.getDoubleExtra(MapsActivity.KEY_STARTING_LAT, 0.0f);
+        double startingLng = actualIntent.getDoubleExtra(MapsActivity.KEY_STARTING_LNG, 0.0f);
 
         assertEquals(45.49701f, startingLat, 0.0001f);
         assertEquals(-73.57877f, startingLng, 0.0001f);
@@ -67,8 +67,8 @@ public class MainActivityEndToEndTest {
         assertEquals("minicap.concordia.campusnav.screens.MapsActivity", actualIntent.getComponent().getClassName());
 
         // Check that the extras contain the correct latitude and longitude USING TOLERANCE AS PRECISION HERE
-        float startingLat = actualIntent.getFloatExtra(MapsActivity.KEY_STARTING_LAT, 0.0f);
-        float startingLng = actualIntent.getFloatExtra(MapsActivity.KEY_STARTING_LNG, 0.0f);
+        double startingLat = actualIntent.getDoubleExtra(MapsActivity.KEY_STARTING_LAT, 0.0f);
+        double startingLng = actualIntent.getDoubleExtra(MapsActivity.KEY_STARTING_LNG, 0.0f);
 
         assertEquals(45.45863f, startingLat, 0.0001f);
         assertEquals(-73.64188f, startingLng, 0.0001f);
